# src/main/java/io/prozy/myfinance/dto/LoginRequestDto.java
package io.prozy.myfinance.dto;

import jakarta.validation.constraints.NotBlank;

public record LoginRequestDto(
        @NotBlank(message = "Login is required")
        String login,

        @NotBlank(message = "Password is required")
        String password
) {}